Extract AccountSubnavigation into a component
