+++ src/app/pages/AccountPage/index.tsx
@@ -11,17 +11,12 @@ import { selectStaking } from 'app/state/staking/selectors'
 import { selectTransaction } from 'app/state/transaction/selectors'
 import { Box } from 'grommet/es6/components/Box'
 import { Layer } from 'grommet/es6/components/Layer'
-import { Nav } from 'grommet/es6/components/Nav'
 import { Spinner } from 'grommet/es6/components/Spinner'
-import { ResponsiveContext } from 'grommet/es6/contexts/ResponsiveContext'
 import { Text } from 'grommet/es6/components/Text'
-import * as React from 'react'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
-import { NavLink, Outlet, useParams } from 'react-router-dom'
-import styled from 'styled-components'
-import { normalizeColor } from 'grommet/es6/utils'
+import { Outlet, useParams } from 'react-router-dom'
 
 import { accountActions } from 'app/state/account'
 import { selectAccount } from 'app/state/account/selectors'
@@ -30,45 +25,12 @@ import { selectActiveWallet, selectHasAccounts, selectHasOneAccount } from 'app/
 import { walletActions } from 'app/state/wallet'
 import { AccountSummary } from './Features/AccountSummary'
 import { AccountPageParams } from './validateAccountPageRoute'
-import { Button } from 'grommet/es6/components/Button'
-
-const StyledNavItem = styled(NavLink)`
-  padding: ${({ theme }) => theme.global?.edgeSize?.small};
-
-  :hover {
-    background-color: ${({ theme }) => normalizeColor('background-contrast', theme)};
-  }
-
-  &.active {
-    background-color: ${({ theme }) => normalizeColor('background-back', theme)};
-  }
-
-  @media only screen and (max-width: ${({ theme }) => `${theme.global?.breakpoints?.small?.value}px`}) {
-    padding: ${({ theme }) => theme.global?.edgeSize?.xsmall};
-  }
-`
-
-interface NavItemProps {
-  counter?: number
-  label: string
-  route: string
-}
-
-const NavItem = ({ counter, label, route }: NavItemProps) => {
-  return (
-    <StyledNavItem end to={route}>
-      <Button badge={counter} tabIndex={-1}>
-        {label}
-      </Button>
-    </StyledNavItem>
-  )
-}
+import { AccountSubnavigation } from './Features/AccountSubnavigation'
 
 interface AccountPageProps {}
 
 export function AccountPage(props: AccountPageProps) {
   const { t } = useTranslation()
-  const isMobile = React.useContext(ResponsiveContext) === 'small'
   const address = useParams<keyof AccountPageParams>().address!
   const dispatch = useDispatch()
 
@@ -145,39 +107,8 @@ export function AccountPage(props: AccountPageProps) {
             wallet={wallet}
             walletHasAccounts={walletHasAccounts}
           />
-          <Nav
-            background="background-front"
-            justify={isMobile ? 'evenly' : 'start'}
-            margin={{ vertical: 'small' }}
-            direction="row"
-            gap="none"
-            wrap
-          >
-            <NavItem
-              label={t('account.subnavigation.transactions', 'Transactions')}
-              route={`/account/${address}`}
-            />
-
-            <NavItem
-              counter={stake.delegations?.length}
-              label={
-                isMobile
-                  ? t('account.subnavigation.mobileActiveDelegations', 'Delegations')
-                  : t('account.subnavigation.activeDelegations', 'Active delegations')
-              }
-              route="active-delegations"
-            />
-
-            <NavItem
-              counter={stake.debondingDelegations?.length}
-              label={
-                isMobile
-                  ? t('account.subnavigation.mobileDebondingDelegations', 'Debonding')
-                  : t('account.subnavigation.debondingDelegations', 'Debonding delegations')
-              }
-              route="debonding-delegations"
-            />
-          </Nav>
+          <Box margin={{ bottom: 'small' }}></Box>
+          <AccountSubnavigation />
           <Outlet />
         </>
       )}
